Create multiple files for the package Issue: #54
