+++ .travis.yml
@@ -1,7 +1,7 @@
 dist: xenial   # required for Python >= 3.7
 language: python
 python:
-  - "2.7"
+  - "3.0"
   - "3.6"
   - "3.7"
 install:
