+++ onebuild/input_parser.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+
+
+def command_to_run(arguments):
+    return "help" if len(arguments) == 1 else arguments[1]
