+++ tests/test_1build.py
@@ -1,13 +1,14 @@
 #!/usr/bin/env python
 
 import imp
-from test_utils import dash
 
+from .test_utils import dash
+from onebuild.main import run
 build = imp.load_source('1build', '1build')
 
 
 def test_build_successful_command(capsys):
-    build.run("tests/data/build_file.yaml", ['file_name', 'build'])
+    run("tests/data/build_file.yaml", ['file_name', 'build'])
     captured = capsys.readouterr()
     expected_message = "" + dash + "\n" \
                                    "Name: build\n" \
@@ -19,7 +20,7 @@ def test_build_successful_command(capsys):
 
 
 def test_should_fail_with_invalid_file_message_if_file_is_not_in_correct_yaml_format(capsys):
-    build.run("tests/data/invalid_yaml_file.yaml", ['file_name', 'build'])
+    run("tests/data/invalid_yaml_file.yaml", ['file_name', 'build'])
     captured = capsys.readouterr()
 
     invalid_file_error_message = "Error in parsing 'tests/data/invalid_yaml_file.yaml' config file. Make sure file is in correct format.\n" \
@@ -35,7 +36,7 @@ def test_should_fail_with_invalid_file_message_if_file_is_not_in_correct_yaml_fo
 
 
 def test_should_print_help_on_help_command(capsys):
-    build.run("tests/data/build_file.yaml", ['file_name', 'help'])
+    run("tests/data/build_file.yaml", ['file_name', 'help'])
     captured = capsys.readouterr()
 
     invalid_file_error_message = "Usage: 1build <command_name> \n\n" \
@@ -48,7 +49,7 @@ def test_should_print_help_on_help_command(capsys):
 
 
 def test_should_print_help_if_no_command_specified(capsys):
-    build.run("tests/data/build_file.yaml", ['file_name'])
+    run("tests/data/build_file.yaml", ['file_name'])
     captured = capsys.readouterr()
 
     invalid_file_error_message = "Usage: 1build <command_name> \n\n" \
@@ -61,10 +62,10 @@ def test_should_print_help_if_no_command_specified(capsys):
 
 
 def test_should_print_command_not_found_if_no_command_found_with_given_name(capsys):
-    build.run("tests/data/build_file.yaml", ['file_name', 'random'])
+    run("tests/data/build_file.yaml", ['file_name', 'random'])
     captured = capsys.readouterr()
 
-    invalid_file_error_message = "No command 'random' found in config file 'tests/data/build_file.yaml'\n\n" \
+    invalid_file_error_message = "No command 'random' found in config file\n\n" \
                                  "Usage: 1build <command_name> \n\n" \
                                  "project: Sample Project\n" + \
                                  "commands:\n" + \
